requirements.txt
streamlit>=1.30
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9
trafilatura>=1.6
readability-lxml>=0.8
html2text>=2024.2

langchain>=0.1.15
langchain-community>=0.0.30
langchain-openai>=0.0.8
langchain-core>=0.1.33

openai>=1.6
python-dotenv>=1.0
pydantic>=2.5
tqdm>=4.66
